requirements.txt
pytubefix
youtube-transcript-api
ipykernel
groq
python-dotenv
google-genai
chromadb
streamlit
pysqlite3-binary
nltk
pydub
audioop-lts
flake8
